animated login page

+++ app/dashboard.js
@@ -1,6 +1,7 @@
 //retrieve the email stored in localstorage
 const email = localStorage.getItem("email");
-
+const loadingAnime = document.getElementById("loadingAnimation");
+let data;
 auth.onAuthStateChanged((user) => {
   //if the user has signed in, go ahead and get the user data from the database
   if (user) {
@@ -11,7 +12,9 @@ auth.onAuthStateChanged((user) => {
       .get()
       .then((querySnapshot) => {
         querySnapshot.forEach((doc) => {
-          const data = doc.data();
+          data = doc.data();
+          console.log(data);
+
           //get their ids and insert the data into them
           document.getElementById("dob").innerHTML = data.dob;
           document.getElementById("firstName1").innerHTML = data.firstname;
@@ -43,17 +46,40 @@ auth.onAuthStateChanged((user) => {
           document.getElementById("programme").innerHTML = data.programme;
           document.getElementById("resultSlip").innerHTML = data.resultslip;
           document.getElementById("resultSlip").innerHTML = data.resultslip;
+
+          /*making perfect circle for animation and 
+          removing animation from the page */
+
+          if (data !== null) {
+            setTimeout(() => {
+              document.getElementById("loading").style.borderRight =
+                "5px solid white";
+            }, 100);
+            setTimeout(() => {
+              document.getElementById("loading").style.borderBottom =
+                "5px solid white";
+            }, 200);
+          }
+          setTimeout(() => {
+            loadingAnime.classList.add("display-active");
+          }, 300);
         });
       });
   }
 });
+
+/**signing users out of the dashboard, clearing
+ * localStorage and redirecting the back to the homepage
+ */
 signout.addEventListener("click", (e) => {
-    e.preventDefault()
+  e.preventDefault();
   auth
     .signOut()
-    .then(() => {
-      localStorage.clear();
-      location.href = '/src/index.html'
+    .then((user) => {
+      if (user == null) {
+        localStorage.clear();
+        location.href = "/src/index.html";
+      }
     })
     .catch((error) => {
       console.log(error);
